requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bidict==0.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
dnspython==2.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
iniconfig==2.0.0
minio==7.2.10
numpy==2.1.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pika==1.3.2
pluggy==1.5.0
pycparser==2.22
pycryptodome==3.21.0
pymongo==4.10.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
requests==2.32.3
simple-websocket==1.1.0
six==1.16.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
